Add note about build-docker-image job going red
There's a dependency in the `build-docker-image` CI job that requires
the blobs. When bumping to a new minor line of Golang, the blobs will be
missing.

We considered adding a pass constraint on the commit having passed
`bump` first to ensure the blobs are there. However, there may be
circumstances where there's a non-bump change to the release that
requires the docker image to be rebuilt.

Signed-off-by: Aram Price <pricear@vmware.com>