Add task to check for fixed CVEs between versions
This script scans for vulnerabilities on the current version of
input_repo, then checks out the previous released version and checks
again. It will exit non-zero if there are no fixed vulnerabilities
between the two versions.

It produces a list of the fixed CVEs, raw JSON output from the scanning
tool, and a pre-built release notes with CVE IDs + title / description.

[#183168350] Periodically create releases of "bosh releases" so that auto-bumped dependencies are available for consumption

Signed-off-by: Brian Upton <bupton@vmware.com>