Make `golang-*-linux` pkgs respect prefixed name
When vendoring another bosh package with a PREFIX (e.g. `bosh
vendor-package PACKAGE --prefix PREFIX`) that depends one of the `golang-*-linux`
packages, the `golang-*-linux` will be vendored with prefix (e.g.
`PREFIX-golang-*-linux`). This commit changes the assumption that the
`golang-*-linux` is always named `golang-*-linux` and instead injects
the name of the package into the compilation and runtime environment
scripts to respect the prefix.