Consolidate windows env scripts; inject pkg name
This commit modifies the packaging script of the windows packages to
calculate the name of the package (in case it's been vendored and
renamed) and inject that name into the compilation and runtime
environment scripts.

This change also allows all of different versions of golang windows
packages to use the same base scripts, consolidating them.