🐞 Publish release with AWS's assume_role
The Docker image we were using had a too-old version of the BOSH CLI,
one which didn't have the ability to use `assume_role` when uploading
blobs. In fact, the Docker image we had been using has been deprecated
in favor of the `bosh/integration` image.

We took the opportunity to use the more-lean `bosh/cli` and
`bosh/warden-cpi` images where possible. Also, where possible, we
updated the Concourse resource to the `registry-image` rather than the
old `docker-image`.

Fixes, when bumping Golang:
```
Uploading blobs:
  Creating blob for path 'go1.19.9.linux-amd64.tar.gz':
    Creating blob in inner blobstore:
      Generating blobstore ID:
        upload failure: AccessDenied: Access Denied
	status code: 403, ...
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>