Add the ability to extract package version
This script will add the ability to extract package version by
passing in the SHA for spec.lock. Users of this release often time have
a hard time telling which version of GO is used under the release. Teams
have came up with creative solutions (e.g. adding a document when
bumping this package), but those solution can fall out of date when a
commit is reverted. This script will print out the GO_VERION used within
that package.