No longer `go mod tidy` with previous go version
At some point in the past, perhaps when dealing with non-go-modded
repos, we needed to do this.

We also no longer need to specify the current go minor when running `go
mod tidy` since we've already run `go mod edit -go=1.$NEW_GO_MINOR`,
which changes the version in the `go.mod` file.

Co-authored-by: Aram Price <pricear@vmware.com>