fix: always use the local tool-chain
Starting with go1.21 the go command will automatically download the
appropriate version of the go tool-chain in case the currently executing
tool-chain is not compatible with the module that is being built.

Since we build releases as self-contained packages that should not
require a internet connection to function this commit disables that
behaviour by setting `GOTOOLCHAIN=local` in the compile env. If the
tool-chain is unable to compile the module, it will fail as it was the
case with pre-go1.21 versions.

See: https://go.dev/doc/toolchain