Enable the CI task `bump-golang-package` to remove packages
This change makes it possible to specify packages to be removed when
bumping a golang package. The removal happens before packages are added
to guard against a case where `PACKAGES_TO_REMOVE` and
`PACKAGES_TO_REMOVE` have the same content.

This can be useful if a release has `golang-1.X-linux` installed and
wishes to replace that package with a newer minor as in
`golang-1.Y-linux`. In this case `PACKAGES_TO_REMOVE` would be an array
containing to package to remove (`golang-1.X-linux`).

Commit message includes a list of removed packages.