Update bump-deps to handle odd toolchain behavior
go get now treats golang and the toolchain as dependencies when performing a "go get -u ./..."

go get is also the new desired way to modify the go and toolchain declarations in the mod file

Because of this, we've removed the "go mod edit" since the go version was then being modified
when we ran "go get -u" with no go declaration.

We now just perfrom a "go get -u" with the go mod version to update to the desired version.

go mod tidy also requires the desired go version otherwise it will update to the version
found in the local toolchain if it is greater.

Updated the param from NEW_GO_MINOR to DESIRED_GO_MAJOR_MINOR to make it more clear what it
does, as well as allowing users to specify "1.21" rather than just "21".

Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>