Correct commit message for bump deps script
There was a bug in our detection of a go version change - we were
comparing the go version string with a patch with one with just a
major.minor version, so it was always saying there was a go upgrade.

To fix this, we strip the patch version from the LHS of the comparison.

Co-authored-by: Chris Selzo <chris.selzo@broadcom.com>