Touch sentinel file when release condition is met
Previously, the shared scripts that check whether a release condition
has been met relied on exiting 0 to indicate a release is needed. While
this works when checking a single condition, some releases (such as the
bosh-cli) need to check multiple conditions (e.g. CVEs have been
patched, OR a newer golang is being used to compile it).

Now, each script will touch a `release-notes/needs-release` sentinel
file when the condition for the script is met. This should allow
multiple check scripts to be chained together, which will also combine
the release notes generated by each.