CI: bump-deps handles multiple GOOS's
Some projects support multiple OS targets, this change allows users of
this task to pass a comman-separated list of valid `GOOS` values as
`GOOS_LIST` to the task, and the Golang update command will be run once
per value in `GOOS_LIST`. This value defaults to `linux`

This resolves an issue where non-linux OS specific packages were not
being updated.

Signed-off-by: Julian Hjortshoj <julian.hjortshoj@broadcom.com>